refactor: normalize state subtracting L on generate instead of code

+++ source/Lib/DecLib/ANSDecoder.h
@@ -11,7 +11,7 @@ class ANSDecoder {
 
     Context context;
     ContextModeler contextModeler = ContextModeler();
-    uint16_t contextualizedStates[2][13] = {};
+    uint8_t contextualizedStates[2][13] = {};
     uint8_t equiprobableState = 0;
 
     TensorBitwidth tensorBitwidth = {};
